remove some more code duplication from connection classes
